Move data generation utilities to utilities.py
