+++ demo/demo_1_shepp_logan.py
@@ -27,6 +27,7 @@ import numpy as np
 import time
 import pprint
 import mbirjax as mj
+import mbirjax.utilities
 
 """**Set the geometry parameters**"""
 
@@ -42,9 +43,9 @@ num_det_channels = 128
 
 # Generate simulated data
 # In a real application you would not have the phantom, but we include it here for later display purposes
-phantom, sinogram, params = mj.generate_demo_data(object_type=object_type, model_type=model_type,
-                                                  num_views=num_views, num_det_rows=num_det_rows,
-                                                  num_det_channels=num_det_channels)
+phantom, sinogram, params = mbirjax.utilities.generate_demo_data(object_type=object_type, model_type=model_type,
+                                                                 num_views=num_views, num_det_rows=num_det_rows,
+                                                                 num_det_channels=num_det_channels)
 angles = params['angles']
 
 # View the sinogram
