+++ experiments/coimg-2025/spectral_response.py
@@ -6,6 +6,8 @@ import mbirjax.parallel_beam
 from scipy.sparse.linalg import svds, eigsh, aslinearoperator, LinearOperator
 import jax
 
+import mbirjax.utilities
+
 if __name__ == "__main__":
     """
     This is a script to develop, debug, and tune the vcd reconstruction with a parallel beam projector
@@ -29,7 +31,7 @@ if __name__ == "__main__":
         # Generate phantom
         recon_shape = parallel_model.get_params('recon_shape')
         num_recon_rows, num_recon_cols, num_recon_slices = recon_shape[:3]
-        phantom = mbirjax.generate_3d_shepp_logan_low_dynamic_range((num_det_channels,num_det_channels,num_det_channels))
+        phantom = mbirjax.utilities.generate_3d_shepp_logan_low_dynamic_range((num_det_channels, num_det_channels, num_det_channels))
         phantom = phantom[:, :, num_det_channels // 2]
 
         # Generate indices of pixels
