+++ experiments/par_cone/demo_2_microscopy.py
@@ -3,6 +3,7 @@ import pprint
 import jax.numpy as jnp
 import mbirjax as mj
 import mbirjax.preprocess as mjp
+import mbirjax.utilities
 
 if __name__ == "__main__":
     """
@@ -39,7 +40,7 @@ if __name__ == "__main__":
     # Generate 3D Shepp Logan phantom
     print('Creating phantom')
     recon_shape = ct_model.get_params("recon_shape")
-    phantom = mj.generate_3d_shepp_logan_low_dynamic_range(recon_shape)
+    phantom = mbirjax.utilities.generate_3d_shepp_logan_low_dynamic_range(recon_shape)
 
     # Generate synthetic sinogram data
     sinogram = ct_model.forward_project(phantom)
