+++ experiments/sandboxes/cone_beam_dev.py
@@ -11,6 +11,8 @@ import mbirjax
 import jax.numpy as jnp
 import jax
 
+import mbirjax.utilities
+
 if __name__ == "__main__":
 
     sino = jnp.zeros((512, 1000, 1536))
@@ -84,7 +86,7 @@ if __name__ == "__main__":
     recon_shape = conebeam_model.get_params('recon_shape')
     num_recon_rows, num_recon_cols, num_recon_slices = recon_shape[:3]
     with jax.default_device(main_device):
-        phantom = mbirjax.gen_cube_phantom(recon_shape)
+        phantom = mbirjax.utilities.gen_cube_phantom(recon_shape)
 
         # Generate indices of pixels and sinogram
         full_indices = mbirjax.gen_full_indices(recon_shape)
