+++ experiments/timing_and_memory/cumulative_memory.py
@@ -4,6 +4,7 @@ import matplotlib.gridspec as gridspec
 import jax.numpy as jnp
 import time
 import mbirjax
+import mbirjax.utilities
 
 
 def create_tiled_heatmap(plot_values, outer_labels, outer_values, inner_labels, inner_values,
@@ -167,7 +168,7 @@ if __name__ == "__main__":
 
                     # Generate phantom for forward projection
                     recon_shape = parallel_model.get_params('recon_shape')
-                    phantom = mbirjax.gen_cube_phantom(recon_shape)
+                    phantom = mbirjax.utilities.gen_cube_phantom(recon_shape)
 
                     # Get a subset of the given size
                     indices = np.arange(ni, dtype=int)
