+++ experiments/translation/exp_translation.py
@@ -1,4 +1,5 @@
 import mbirjax as mj
+import mbirjax.utilities
 
 
 def main():
@@ -22,7 +23,7 @@ def main():
     words = ["Purdue", "Presents", "Translation", "Tomography"] # List of words to render in the text phantom
 
     # Generate translation vectors
-    translation_vectors = mj.gen_translation_vectors(num_x_translations, num_z_translations, x_spacing, z_spacing)
+    translation_vectors = mbirjax.utilities.gen_translation_vectors(num_x_translations, num_z_translations, x_spacing, z_spacing)
 
     # Compute sinogram shape
     sino_shape = (translation_vectors.shape[0], num_det_rows, num_det_channels)
@@ -38,7 +39,7 @@ def main():
     mj.display_translation_vectors(translation_vectors, recon_shape)
 
     # Generate ground truth phantom
-    gt_recon = mj.gen_translation_phantom(recon_shape=recon_shape, option=phantom_type, words=words)
+    gt_recon = mbirjax.utilities.gen_translation_phantom(recon_shape=recon_shape, option=phantom_type, words=words)
 
     # View test sample
     mj.slice_viewer(gt_recon.transpose(0, 2, 1), title='Ground Truth Recon', slice_label='View', slice_axis=0)
