+++ mbirjax/utilities.py
@@ -1,9 +1,14 @@
 import os, sys
+from enum import Enum
+from typing import Union
 
+import jax
 # === Core scientific/plotting libraries ===
 import matplotlib.pyplot as plt
 import numpy as np
 import jax.numpy as jnp
+from PIL import ImageFont, Image, ImageDraw
+from jax import numpy as jnp
 
 # === Project-specific imports ===
 import mbirjax as mj
@@ -600,3 +605,450 @@ def import_recon_hdf5(file_path):
 
     return recon, recon_dict
 
+
+def generate_3d_shepp_logan_reference(phantom_shape):
+    """
+    Generate a 3D Shepp Logan phantom based on below reference.
+
+    Kak AC, Slaney M. Principles of computerized tomographic imaging. Page.102. IEEE Press, New York, 1988. https://engineering.purdue.edu/~malcolm/pct/CTI_Ch03.pdf
+
+    Args:
+        phantom_shape (tuple or list of ints): num_rows, num_cols, num_slices
+
+    Return:
+        out_image: 3D array, num_slices*num_rows*num_cols
+
+    Note:
+        This function produces 6 intermediate arrays that each have shape phantom_shape, so if phantom_shape is
+        large, then this will use a lot of peak memory.
+    """
+
+    # The function describing the phantom is defined as the sum of 10 ellipsoids inside a 2×2×2 cube:
+    sl3d_paras = [
+        {'x0': 0.0, 'y0': 0.0, 'z0': 0.0, 'a': 0.69, 'b': 0.92, 'c': 0.9, 'gamma': 0, 'gray_level': 2.0},
+        {'x0': 0.0, 'y0': 0.0, 'z0': 0.0, 'a': 0.6624, 'b': 0.874, 'c': 0.88, 'gamma': 0, 'gray_level': -0.98},
+        {'x0': -0.22, 'y0': 0.0, 'z0': -0.25, 'a': 0.41, 'b': 0.16, 'c': 0.21, 'gamma': 108, 'gray_level': -0.02},
+        {'x0': 0.22, 'y0': 0.0, 'z0': -0.25, 'a': 0.31, 'b': 0.11, 'c': 0.22, 'gamma': 72, 'gray_level': -0.02},
+        {'x0': 0.0, 'y0': 0.35, 'z0': -0.25, 'a': 0.21, 'b': 0.25, 'c': 0.5, 'gamma': 0, 'gray_level': 0.02},
+        {'x0': 0.0, 'y0': 0.1, 'z0': -0.25, 'a': 0.046, 'b': 0.046, 'c': 0.046, 'gamma': 0, 'gray_level': 0.02},
+        {'x0': -0.08, 'y0': -0.65, 'z0': -0.25, 'a': 0.046, 'b': 0.023, 'c': 0.02, 'gamma': 0, 'gray_level': 0.01},
+        {'x0': 0.06, 'y0': -0.65, 'z0': -0.25, 'a': 0.046, 'b': 0.023, 'c': 0.02, 'gamma': 90, 'gray_level': 0.01},
+        {'x0': 0.06, 'y0': -0.105, 'z0': 0.625, 'a': 0.056, 'b': 0.04, 'c': 0.1, 'gamma': 90, 'gray_level': 0.02},
+        {'x0': 0.0, 'y0': 0.1, 'z0': 0.625, 'a': 0.056, 'b': 0.056, 'c': 0.1, 'gamma': 0, 'gray_level': -0.02}
+    ]
+
+    num_rows, num_cols, num_slices = phantom_shape
+    axis_x = np.linspace(-1.0, 1.0, num_cols)
+    axis_y = np.linspace(1.0, -1.0, num_rows)
+    axis_z = np.linspace(-1.0, 1.0, num_slices)
+
+    x_grid, y_grid, z_grid = np.meshgrid(axis_x, axis_y, axis_z)
+    image = x_grid * 0.0
+
+    for el_paras in sl3d_paras:
+        image += _gen_ellipsoid(x_grid=x_grid, y_grid=y_grid, z_grid=z_grid, x0=el_paras['x0'], y0=el_paras['y0'],
+                               z0=el_paras['z0'],
+                               a=el_paras['a'], b=el_paras['b'], c=el_paras['c'],
+                               gamma=el_paras['gamma'] / 180.0 * np.pi,
+                               gray_level=el_paras['gray_level'])
+
+    return image.transpose((1, 0, 2))
+
+
+def generate_3d_shepp_logan_low_dynamic_range(phantom_shape, device=None):
+    """
+    Generates a 3D Shepp-Logan phantom with specified dimensions.
+
+    Args:
+        phantom_shape (tuple): Phantom shape in (rows, columns, slices).
+        device (jax device): Device on which to place the output phantom.
+
+    Returns:
+        ndarray: A 3D numpy array of shape phantom_shape representing the voxel intensities of the phantom.
+
+    Note:
+        This function uses a memory-efficient approach to generating large phantoms.
+    """
+    # Get space for the result and set up the grids for add_ellipsoid
+    with jax.default_device(device):
+        phantom = jnp.zeros(phantom_shape, device=device)
+    N, M, P = phantom_shape
+    x_locations = jnp.linspace(-1, 1, N)
+    y_locations = jnp.linspace(-1, 1, M)
+    z_locations = jnp.linspace(-1, 1, P)
+    x_grid, y_grid = jnp.meshgrid(x_locations, y_locations, indexing='ij')
+    i_grid, j_grid = jnp.meshgrid(jnp.arange(N), jnp.arange(M), indexing='ij')
+    grids = (x_grid, y_grid, i_grid, j_grid)
+
+    # Main ellipsoid
+    phantom = add_ellipsoid(phantom, grids, z_locations, 0, 0, 0, 0.69, 0.92, 0.9, intensity=1)
+    # Smaller ellipsoids and other structures
+    phantom = add_ellipsoid(phantom, grids, z_locations, 0, 0.0184, 0, 0.6624, 0.874, 0.88, intensity=-0.8)
+    phantom = add_ellipsoid(phantom, grids, z_locations, 0.22, 0, 0, 0.41, 0.16, 0.21, angle=108, intensity=-0.2)
+    phantom = add_ellipsoid(phantom, grids, z_locations, -0.22, 0, 0, 0.31, 0.11, 0.22, angle=72, intensity=-0.2)
+    phantom = add_ellipsoid(phantom, grids, z_locations, 0, 0.35, 0, 0.21, 0.25, 0.5, intensity=0.1)
+    phantom = add_ellipsoid(phantom, grids, z_locations, 0, 0.1, 0, 0.046, 0.046, 0.046, intensity=0.1)
+    phantom = add_ellipsoid(phantom, grids, z_locations, 0, -0.1, 0, 0.046, 0.046, 0.046, intensity=0.1)
+    phantom = add_ellipsoid(phantom, grids, z_locations, -0.08, -0.605, 0, 0.046, 0.023, 0.02, angle=0, intensity=0.1)
+    phantom = add_ellipsoid(phantom, grids, z_locations, 0, -0.605, 0, 0.023, 0.023, 0.02, angle=0, intensity=0.1)
+
+    return phantom
+
+
+def gen_translation_phantom(recon_shape, option, words, fill_rate=0.05, font_size=20):
+    """
+    Generate a synthetic ground truth phantom based on the selected option.
+
+    Args:
+        option (str): Phantom type to generate. Options are 'dots' or 'text'.
+        recon_shape (tuple[int, int, int]): Shape of the reconstruction volume.
+        words (list[str]): List of ASCII words to render.
+        fill_rate (float, optional): Fill rate of the reconstruction volume. Default is 0.05.
+        font_size (int, optional): Font size of the ASCII words. Default is 20.
+
+    Returns:
+        np.ndarray: Generated phantom volume.
+    """
+    if option == 'dots':
+        return gen_dot_phantom(recon_shape, fill_rate)
+    elif option == 'text':
+        return gen_text_phantom(recon_shape, words, font_size)
+    else:
+        raise ValueError(f"Unsupported phantom option: {option}")
+
+
+def gen_dot_phantom(recon_shape, fill_rate):
+    """
+    Generate a synthetic ground truth reconstruction volume.
+
+    Args:
+        recon_shape (tuple[int, int, int]): Shape of the reconstruction volume.
+        fill_rate (float): Fill rate of the reconstruction volume.
+
+    Returns:
+        np.ndarray: Ground truth reconstruction volume with sparse binary features.
+    """
+    np.random.seed(42)
+    gt_recon = np.zeros(recon_shape, dtype=np.float32)
+
+    y_pad = recon_shape[0] // 6
+    central_start = y_pad
+    central_end = recon_shape[0] - y_pad
+
+    row_size = recon_shape[1] * recon_shape[2]
+    num_ones_per_row = int(row_size * fill_rate)
+
+    for row_idx in range(central_start, central_end):
+        flat_row = gt_recon[row_idx].flatten()
+        positions_ones = np.random.choice(row_size, num_ones_per_row, replace=False)
+        flat_row[positions_ones] = 1.0
+        gt_recon[row_idx] = flat_row.reshape(recon_shape[1:])
+
+    return gt_recon
+
+
+def gen_text_phantom(recon_shape, words, font_size, font_path="DejaVuSans.ttf"):
+    """
+    Generate a 3D text phantom with binary word patterns embedded in specific slices.
+
+    Args:
+        recon_shape (tuple[int, int, int]): Shape of the phantom volume (num_rows, num_cols, num_slices).
+        words (list[str]): List of ASCII words to render.
+        font_size (int): Font size of ASCII words.
+        font_path (str, optional): Path to the TrueType font file. Default is "DejaVuSans.ttf".
+
+    Returns:
+        np.ndarray: A 3D numpy array of shape `recon_shape` containing the text phantom.
+    """
+    positions = []
+    for i in range(1, len(words) + 1):
+        positions.append((recon_shape[0] // (len(words) + 1) * i, recon_shape[1] // 2, recon_shape[2] // 2))
+
+    array_size = np.minimum(recon_shape[1], recon_shape[2])
+
+    phantom = np.zeros(recon_shape, dtype=np.float32)
+    try:
+        font = ImageFont.truetype(font_path, size=font_size)
+    except OSError:
+        from pathlib import Path
+        fallback_paths = [
+            "/System/Library/Fonts/Supplemental/Arial.ttf",  # macOS fallback
+            "/Library/Fonts/Arial.ttf",  # Additional macOS path
+            "/usr/share/fonts/truetype/dejavu/DejaVuSans.ttf",  # Linux fallback
+        ]
+        for fallback in fallback_paths:
+            if Path(fallback).exists():
+                font = ImageFont.truetype(fallback, size=font_size)
+                break
+        else:
+            raise FileNotFoundError(
+                f"Could not find a usable font. Tried the following paths:\n"
+                + "\n".join(fallback_paths)
+                + "\nPlease install one of these fonts or specify a valid font_path."
+            )
+
+    for word, (r, c, s) in zip(words, positions):
+        img = Image.new('L', (array_size, array_size), 0)
+        draw = ImageDraw.Draw(img)
+
+        text_box = draw.textbbox((0, 0), word, font=font)
+        text_width = text_box[2] - text_box[0]
+        text_height = text_box[3] - text_box[1]
+
+        x = (array_size - text_width) // 2 - text_box[0]
+        y = (array_size - text_height) // 2 - text_box[1]
+        draw.text((x, y), word, fill=1, font=font)
+
+        word_array = np.array(img.rotate(-90, expand=True).transpose(Image.FLIP_LEFT_RIGHT))
+        word_array = (word_array > 0).astype(np.float32)
+
+        # Crop or pad word_array to fit in the recon volume
+        r_start = max(0, r)
+        r_end = min(recon_shape[0], r + 1)
+        c_start = max(0, c - array_size // 2)
+        c_end = min(recon_shape[1], c_start + array_size)
+        s_start = max(0, s - array_size // 2)
+        s_end = min(recon_shape[2], s_start + array_size)
+
+        word_crop = word_array[:(c_end - c_start), :(s_end - s_start)]
+        phantom[r_start:r_end, c_start:c_end, s_start:s_end] = word_crop
+
+    return phantom
+
+
+def _gen_ellipsoid(x_grid, y_grid, z_grid, x0, y0, z0, a, b, c, gray_level, alpha=0, beta=0, gamma=0):
+    """
+    Return an image with a 3D ellipsoid in a 3D plane with a center of [x0,y0,z0] and ...
+
+    Args:
+        x_grid(jax array): 3D grid of X coordinate values.
+        y_grid(jax array): 3D grid of Y coordinate values.
+        z_grid(jax array): 3D grid of Z coordinate values.
+        x0(float): horizontal center of ellipsoid.
+        y0(float): vertical center of ellipsoid.
+        z0(float): normal center of ellipsoid.
+        a(float): X-axis radius.
+        b(float): Y-axis radius.
+        c(float): Z-axis radius.
+        gray_level(float): Gray level for the ellipse.
+        alpha(float): [Default=0.0] counter-clockwise angle of rotation by X-axis in radians.
+        beta(float): [Default=0.0] counter-clockwise angle of rotation by Y-axis in radians.
+        gamma(float): [Default=0.0] counter-clockwise angle of rotation by Z-axis in radians.
+
+    Return:
+        ndarray: 3D array with the same shape as x_grid, y_grid, and z_grid
+
+    """
+    # Generate Rotation Matrix.
+    rx = np.array([[1, 0, 0], [0, np.cos(-alpha), -np.sin(-alpha)], [0, np.sin(-alpha), np.cos(-alpha)]])
+    ry = np.array([[np.cos(-beta), 0, np.sin(-beta)], [0, 1, 0], [-np.sin(-beta), 0, np.cos(-beta)]])
+    rz = np.array([[np.cos(-gamma), -np.sin(-gamma), 0], [np.sin(-gamma), np.cos(-gamma), 0], [0, 0, 1]])
+    r = np.dot(rx, np.dot(ry, rz))
+
+    cor = np.array([x_grid.flatten() - x0, y_grid.flatten() - y0, z_grid.flatten() - z0])
+
+    image = ((np.dot(r[0], cor)) ** 2 / a ** 2 + (np.dot(r[1], cor)) ** 2 / b ** 2 + (
+        np.dot(r[2], cor)) ** 2 / c ** 2 <= 1.0) * gray_level
+
+    return image.reshape(x_grid.shape)
+
+
+@jax.jit
+def add_ellipsoid(current_volume, grids, z_locations, x0, y0, z0, a, b, c, angle=0, intensity=1.0):
+    """
+    Add an ellipsoid to an existing jax array.  This is done using lax.scan over the z slices to avoid
+    using really large arrays when the volume is large.
+
+    Args:
+        current_volume (jax array): 3D volume
+        grids (tuple):  A tuple of x_grid, y_grid, i_grid, j_grid obtained as in generate_3d_shepp_logan_low_dynamic_range
+        z_locations (jax array): A 1D array of z coordinates of the volume
+        x0 (float): x center for the ellipsoid
+        y0 (float): y center for the ellipsoid
+        z0 (float): z center for the ellipsoid
+        a (float): x radius
+        b (float): y radius
+        c (float): z radius
+        angle (float): angle of rotation of the ellipsoid in the xy plane around (x0, y0)
+        intensity (float): The constant value of the ellipsoid to be added.
+
+    Returns:
+        3D jax array: current_volume + ellipsoid
+    """
+
+    # Unpack the grids and determine the xy locations for this angle
+    x_grid, y_grid, i_grid, j_grid = grids
+    cos_angle = jnp.cos(jnp.deg2rad(angle))
+    sin_angle = jnp.sin(jnp.deg2rad(angle))
+    Xr = cos_angle * (x_grid - x0) + sin_angle * (y_grid - y0)
+    Yr = -sin_angle * (x_grid - x0) + cos_angle * (y_grid - y0)
+
+    # Determine which xy locations will be updated for this ellipsoid
+    xy_norm = Xr**2 / a**2 + Yr**2 / b**2
+
+    def add_slice_vmap(volume_slice, z):
+        return volume_slice + intensity * ((xy_norm + (z - z0)**2 / c**2) <= 1).astype(float)
+
+    volume_map = jax.vmap(add_slice_vmap, in_axes=(2, 0), out_axes=2)
+    current_volume = volume_map(current_volume, z_locations)
+
+    return current_volume
+
+
+class ObjectType(str, Enum):
+    SHEPP_LOGAN = 'shepp-logan'
+    CUBE = 'cube'
+
+
+class ModelType(str, Enum):
+    PARALLEL = 'parallel'
+    CONE = 'cone'
+    TRANSLATION = 'translation'
+
+
+def generate_demo_data(
+    object_type: Union[ObjectType, str] = ObjectType.SHEPP_LOGAN,
+    model_type: Union[ModelType, str] = ModelType.CONE,
+    num_views: int = 64,
+    num_det_rows: int = 96,
+    num_det_channels: int = 128,
+    num_x_translations: int = 7,
+    num_z_translations: int = 7,
+    x_spacing: float = 22,
+    z_spacing: float = 22
+) -> (np.ndarray, np.ndarray):
+    """
+    Create a simple object and a sinogram for demonstration purposes.
+
+    This function will create a 3D volume (aka object or phantom) of the specified type, then use the model type and
+    parameters to create a simulated sinogram.  The object type 'shepp-logan' gives a simplified version of the
+    classic Shepp-Logan test phantom, and type 'cube' gives a simple cube object.
+
+    The output sinogram is a 3D numpy array with shape (num_views, num_det_rows, num_det_channels).  Each 2D array
+    sinogram[view_index] is a simulated image from the detector, with num_det_rows indicating the vertical size
+    and num_det_channels representing the horizontal size.
+
+    Args:
+        object_type (str, optional): One of 'shepp-logan' or 'cube'.  Defaults to 'shepp-logan'.
+        model_type (str, optional): One of 'parallel', 'cone', or 'translation'.  Defaults to 'cone'.
+        num_views (int, optional):  Number of views in the output sinogram.  Defaults to 64. Ignored when model_type is 'translation'
+        num_det_rows (int, optional): Number of rows (vertical) in the output sinogram.  Defaults to 40.
+        num_det_channels (int, optional): Number of channels (horizontal) in the output sinogram.  Defaults to 128.
+        num_x_translations (int, optional): Number of horizontal translations for translation mode.  Defaults to 7.
+        num_z_translations (int, optional): Number of vertical translations for translation mode.  Defaults to 7.
+        x_spacing (float, optional): Horizontal spacing between translations in ALU.  Defaults to 22.
+        z_spacing (float, optional): Vertical spacing between translations in ALU.  Defaults to 22.
+
+    Returns:
+        tuple: (object, sinogram, params)
+            - object (np.ndarray): a volume with shape (num_det_channels, num_det_channels, num_det_rows)
+            - sinogram (np.ndarray): a sinogram with shape (num_views, num_det_rows, num_det_channels)
+            - params (dict): a dict containing 'angles' and, if model_type is 'cone', then also 'source_detector_dist' and 'source_iso_dist'
+    """
+    # Coerce types to Enum
+    object_type = ObjectType(object_type)
+    model_type = ModelType(model_type)
+
+    start_angle = -np.pi
+    end_angle = np.pi
+
+    # Initialize model
+
+    if model_type == ModelType.PARALLEL:
+        sinogram_shape = (num_views, num_det_rows, num_det_channels)
+        angles = jnp.linspace(start_angle, end_angle, num_views, endpoint=False)
+        ct_model_for_generation = mj.ParallelBeamModel(sinogram_shape, angles)
+        params = {'angles': angles}
+    elif model_type == ModelType.CONE:
+        # For cone beam geometry, we need to describe the distances source to detector and source to rotation axis.
+        # np.Inf is an allowable value, in which case this is essentially parallel beam
+        source_detector_dist = 4 * num_det_channels
+        source_iso_dist = source_detector_dist
+        sinogram_shape = (num_views, num_det_rows, num_det_channels)
+        angles = jnp.linspace(start_angle, end_angle, num_views, endpoint=False)
+        ct_model_for_generation = mj.ConeBeamModel(sinogram_shape, angles, source_detector_dist=source_detector_dist,
+                                                   source_iso_dist=source_iso_dist)
+        params = {'angles': angles, 'source_detector_dist': source_detector_dist, 'source_iso_dist': source_iso_dist}
+    elif model_type == ModelType.TRANSLATION:
+        source_iso_dist = np.min(num_det_rows, num_det_channels) / 2
+        source_detector_dist = source_iso_dist
+        translation_vectors = gen_translation_vectors(num_x_translations, num_z_translations, x_spacing, z_spacing)
+        num_views = translation_vectors.shape[0]
+        sinogram_shape = (num_views, num_det_rows, num_det_channels)
+        ct_model_for_generation = mj.TranslationModel(sinogram_shape, translation_vectors, source_detector_dist=source_detector_dist,
+                                                      source_iso_dist=source_iso_dist)
+        params = {'translation_vectors': translation_vectors}
+    else:
+        raise ValueError(f'Invalid model type. Expected one of {[m.value for m in ModelType]}, got {model_type}')
+
+    # Generate phantom
+    print('Creating phantom')
+    recon_shape = ct_model_for_generation.get_params('recon_shape')
+    device = ct_model_for_generation.main_device
+    if object_type == ObjectType.SHEPP_LOGAN:
+        phantom = mbirjax.utilities.generate_3d_shepp_logan_low_dynamic_range(recon_shape, device=device)
+    elif object_type == ObjectType.CUBE:
+        phantom = gen_cube_phantom(recon_shape, device=device)
+    else:
+        raise ValueError(f'Invalid object type. Expected one of {[o.value for o in ObjectType]}, got {object_type}')
+
+    # Generate synthetic sinogram data
+    print('Creating sinogram')
+    sinogram = ct_model_for_generation.forward_project(phantom)
+    sinogram = np.asarray(sinogram)
+
+    return phantom, sinogram, params
+
+
+def gen_translation_vectors(num_x_translations, num_z_translations, x_spacing, z_spacing):
+    """
+    Generate translation vectors for lateral (x) and axial (z) displacements.
+
+    Args:
+        num_x_translations (int): Number of x-direction translations
+        num_z_translations (int): Number of z-direction translations
+        x_spacing (float): Spacing between x translations in ALU
+        z_spacing (float): Spacing between z translations in ALU
+
+    Returns:
+        np.ndarray: Array of shape (num_views, 3) with translation vectors [dx, dy, dz]
+    """
+    num_views = num_x_translations * num_z_translations
+    translation_vectors = np.zeros((num_views, 3))
+
+    x_center = (num_x_translations - 1) / 2
+    z_center = (num_z_translations - 1) / 2
+
+    idx = 0
+    for row in range(num_z_translations):
+        for col in range(num_x_translations):
+            dx = (col - x_center) * x_spacing
+            dz = (row - z_center) * z_spacing
+            dy = 0
+            translation_vectors[idx] = [dx, dy, dz]
+            idx += 1
+
+    return translation_vectors
+
+
+def gen_cube_phantom(recon_shape, device=None):
+    """Code to generate a simple phantom """
+    # Compute phantom height and width
+    num_recon_rows, num_recon_cols, num_recon_slices = recon_shape[:3]
+    phantom_rows = num_recon_rows // 4  # Phantom height
+    phantom_cols = num_recon_cols // 4  # Phantom width
+
+    # Allocate phantom memory
+    phantom = np.zeros((num_recon_rows, num_recon_cols, num_recon_slices))
+
+    # Compute start and end locations
+    start_rows = (num_recon_rows - phantom_rows) // 2
+    stop_rows = (num_recon_rows + phantom_rows) // 2
+    start_cols = (num_recon_cols - phantom_cols) // 2
+    stop_cols = (num_recon_cols + phantom_cols) // 2
+    for slice_index in np.arange(num_recon_slices):
+        shift_cols = int(slice_index * phantom_cols / num_recon_slices)
+        phantom[start_rows:stop_rows, (shift_cols + start_cols):(shift_cols + stop_cols), slice_index] = 1.0 / max(
+            phantom_rows, phantom_cols)
+
+    return jnp.array(phantom, device=device)
