+++ tests/test_denoiser.py
@@ -2,6 +2,8 @@ import numpy as np
 import mbirjax as mj
 import unittest
 
+import mbirjax.utilities
+
 
 class TestProx(unittest.TestCase):
     """
@@ -39,7 +41,7 @@ class TestProx(unittest.TestCase):
 
         # Get some noisy data
         recon_shape = (num_det_channels, num_det_channels, num_det_rows)
-        phantom = mj.generate_3d_shepp_logan_low_dynamic_range(recon_shape)
+        phantom = mbirjax.utilities.generate_3d_shepp_logan_low_dynamic_range(recon_shape)
         phantom_noisy = phantom + sigma_noise * np.random.randn(*recon_shape)
 
         denoiser = mj.QGGMRFDenoiser(phantom.shape)
